Split TS generation into files : move header files to the right place
