+++ src/tools/ts-generator/CMakeLists.txt
@@ -1,7 +1,7 @@
 set(SRCS
         main.cpp
-		linksTSgenerator.h
-		tsGenerationOptions.h
+		include/antares/tools/ts-generator/linksTSgenerator.h
+		include/antares/tools/ts-generator/tsGenerationOptions.h
 		tsGenerationOptions.cpp
 		linksTSgenerator.cpp
 )
@@ -17,15 +17,15 @@ INSTALL(EXPORT ${execname}
 
 target_link_libraries(${execname}
                       PRIVATE
-                      Antares::utils
-                      antares-solver-ts-generator
-                      Antares::study
-		      Antares::checks
+						Antares::utils
+						antares-solver-ts-generator
+						Antares::study
+						Antares::checks
 )
 
 target_include_directories(${execname}
         PRIVATE
-        "${CMAKE_CURRENT_SOURCE_DIR}/include"
+			"${CMAKE_CURRENT_SOURCE_DIR}/include"
 )
 
 import_std_libs(${execname})
