+++ src/tools/ts-generator/linksTSgenerator.cpp
@@ -1,5 +1,5 @@
 
-#include "linksTSgenerator.h"
+#include "antares/tools/ts-generator/linksTSgenerator.h"
 #include <yuni/datetime/timestamp.h>
 
 namespace Antares::TSGenerator
